+++ Runtime/Scripts/Service/Services/StartOrderingService.cs
@@ -15,14 +15,19 @@ namespace VirrVarr
         int numRegisteredObjects = 0;
         bool immediateStartEnabled = false;
 
-        /// <summary>
-        /// Constructs a new Start Ordering service.
-        /// </summary>
         public StartOrderingService()
         {
             registeredStartOrderIndexes = new List<StartOrderRegistration>();
             SweepSceneObjects();
         }
+
+        /// <summary>
+        /// Register a behaviour with its start index.
+        /// Note: Start ordering is only respected for scene object.
+        /// If start order has been run and a behaviour registers, it will start immediately.
+        /// </summary>
+        /// <param name="startOrderIndex">The start index of the behaviour</param>
+        /// <param name="behaviour">The behaviour itself</param>
         public void RegisterStartOrderIndexBehaviour( int startOrderIndex, BehaviourBase behaviour )
         {
             int foundIndex = registeredStartOrderIndexes.FindIndex( (StartOrderRegistration reg ) => { return reg.startOrderIndex == startOrderIndex; } );
@@ -52,6 +57,10 @@ namespace VirrVarr
                 RunStartOrder();
             }
         }
+
+        /// <summary>
+        /// Start all registered behaviours by increasing start index
+        /// </summary>
         private void RunStartOrder()
         {
             // Sort the registrations from lower -> higher
@@ -67,6 +76,9 @@ namespace VirrVarr
 
             immediateStartEnabled = true;
         }
+        /// <summary>
+        /// Sweep the scene for all behaviours in it.
+        /// </summary>
         private void SweepSceneObjects()
         {
             var behaviourBaseList = UnityEngine.GameObject.FindObjectsOfType<BehaviourBase>();
@@ -79,6 +91,13 @@ namespace VirrVarr
                 }
             }
         }
+
+        /// <summary>
+        /// Comparer for start order registrations so that we can order them
+        /// </summary>
+        /// <param name="a">Compare reg a</param>
+        /// <param name="b">Compare reg b</param>
+        /// <returns></returns>
         private static int CompareRegistration( StartOrderRegistration a, StartOrderRegistration b )
         {
             if(a.startOrderIndex == b.startOrderIndex)